+++ src/components/AudioPlayer/PlaybackControls.tsx
@@ -2,48 +2,60 @@ import { useSelector } from 'react-redux';
 import { selectIsPlaying } from 'src/redux/slices/AudioPlayer/state';
 import { withStopPropagation } from 'src/utils/event';
 import styles from './PlaybackControls.module.scss';
-import PlayIcon from '../../../public/icons/play-circle-outline.svg';
+import PlayIcon from '../../../public/icons/play-arrow.svg';
 import ForwardIcon from '../../../public/icons/forward_10.svg';
-import PauseIcon from '../../../public/icons/pause-circle-outline.svg';
+import PauseIcon from '../../../public/icons/pause.svg';
 import ReplayIcon from '../../../public/icons/replay_10.svg';
 import ShareIcon from '../../../public/icons/share.svg';
 import RepeatIcon from '../../../public/icons/ic_repeat_24px 1.svg';
-import Button, { ButtonSize } from '../dls/Button/Button';
+import Button, { ButtonShape, ButtonSize, ButtonVariant } from '../dls/Button/Button';
 import { triggerPauseAudio, triggerPlayAudio, triggerSeek } from './EventTriggers';
 
 const PlaybackControls = () => {
   const isPlaying = useSelector(selectIsPlaying);
   return (
     <div className={styles.container}>
-      <Button icon={<RepeatIcon />} size={ButtonSize.Small} desktopSize={ButtonSize.Medium} />
+      <Button variant={ButtonVariant.Ghost} shape={ButtonShape.Circle} size={ButtonSize.Large}>
+        <RepeatIcon />
+      </Button>
       <Button
-        icon={<ReplayIcon />}
-        size={ButtonSize.Small}
-        desktopSize={ButtonSize.Medium}
+        variant={ButtonVariant.Ghost}
+        shape={ButtonShape.Circle}
+        size={ButtonSize.Large}
         onClick={withStopPropagation(() => triggerSeek(-10))}
-      />
+      >
+        <ReplayIcon />
+      </Button>
       {isPlaying ? (
         <Button
-          icon={<PauseIcon />}
-          size={ButtonSize.Small}
-          desktopSize={ButtonSize.Medium}
+          variant={ButtonVariant.Ghost}
+          shape={ButtonShape.Circle}
+          size={ButtonSize.Large}
           onClick={withStopPropagation(triggerPauseAudio)}
-        />
+        >
+          <PauseIcon />
+        </Button>
       ) : (
         <Button
-          icon={<PlayIcon />}
-          size={ButtonSize.Small}
-          desktopSize={ButtonSize.Medium}
+          variant={ButtonVariant.Ghost}
+          shape={ButtonShape.Circle}
+          size={ButtonSize.Large}
           onClick={withStopPropagation(triggerPlayAudio)}
-        />
+        >
+          <PlayIcon />
+        </Button>
       )}
       <Button
-        icon={<ForwardIcon />}
-        size={ButtonSize.Small}
-        desktopSize={ButtonSize.Medium}
+        variant={ButtonVariant.Ghost}
+        shape={ButtonShape.Circle}
+        size={ButtonSize.Large}
         onClick={withStopPropagation(() => triggerSeek(10))}
-      />
-      <Button icon={<ShareIcon />} size={ButtonSize.Small} desktopSize={ButtonSize.Medium} />
+      >
+        <ForwardIcon />
+      </Button>
+      <Button variant={ButtonVariant.Ghost} shape={ButtonShape.Circle} size={ButtonSize.Large}>
+        <ShareIcon />
+      </Button>
     </div>
   );
 };
