+++ src/components/Navbar/SearchDrawer/SearchDrawer.tsx
@@ -3,7 +3,7 @@ import { selectNavbar, setIsSearchDrawerOpen } from 'src/redux/slices/navbar';
 import { useDispatch, useSelector } from 'react-redux';
 import { useRouter } from 'next/router';
 import useTranslation from 'next-translate/useTranslation';
-import Button, { ButtonSize } from 'src/components/dls/Button/Button';
+import Button, { ButtonShape, ButtonVariant } from 'src/components/dls/Button/Button';
 import useElementComputedPropertyValue from 'src/hooks/useElementComputedPropertyValue';
 import { getSearchResults } from 'src/api';
 import { SearchResponse } from 'types/APIResponses';
@@ -102,11 +102,13 @@ const SearchDrawer: React.FC = () => {
         <div className={styles.headerContentContainer}>
           <div className={styles.headerContent}>
             <Button
-              icon={<IconSearch />}
-              size={ButtonSize.Small}
+              shape={ButtonShape.Circle}
+              variant={ButtonVariant.Ghost}
               disabled={!searchQuery}
               href={`/search?query=${searchQuery}`}
-            />
+            >
+              <IconSearch />
+            </Button>
             <div
               className={classNames(styles.searchInputContainer, {
                 [styles.searchInputContainerRTL]: isRTLInput,
@@ -128,7 +130,13 @@ const SearchDrawer: React.FC = () => {
                 </button>
               )}
             </div>
-            <Button icon={<IconClose />} size={ButtonSize.Small} onClick={closeSearchDrawer} />
+            <Button
+              shape={ButtonShape.Circle}
+              variant={ButtonVariant.Ghost}
+              onClick={closeSearchDrawer}
+            >
+              <IconClose />
+            </Button>
           </div>
         </div>
       </div>
