+++ src/components/chapters/ChapterHeader/index.tsx
@@ -1,7 +1,7 @@
 import Bismillah from 'src/components/dls/Bismillah/Bismillah';
 import React from 'react';
 import PlayChapterAudioButton from 'src/components/QuranReader/PlayChapterAudioButton';
-import Button from 'src/components/dls/Button/Button';
+import Button, { ButtonVariant } from 'src/components/dls/Button/Button';
 import ChapterIconContainer from '../ChapterIcon/ChapterIconContainer';
 import styles from './ChapterHeader.module.scss';
 
@@ -23,7 +23,9 @@ const ChapterHeader: React.FC<Props> = ({ chapterId }) => (
     )}
     <div className={styles.actionsContainer}>
       <PlayChapterAudioButton chapterId={Number(chapterId)} />
-      <Button text="Info" href={`/${chapterId}/info`} />
+      <Button variant={ButtonVariant.Ghost} href={`/${chapterId}/info`}>
+        Info
+      </Button>
     </div>
   </div>
 );
