+++ src/components/dls/Pagination/Pagination.module.scss
@@ -11,5 +11,5 @@
 
 .selectedButton {
   background: var(--color-background-faded-grey-scale);
-  border-radius: var(--border-radius-circle);
+  border-radius: var(--border-radius-default);
 }
